pass _func_, _call_, and _decorate_ into generated code using closures not the globals
The decorator module currently copies a functions globals to evaluate the source template in. Having multiple copies of a module's globals around is bad. It can lead to subtle memory leaks. Instead, we should just evaluate the generated function in the wrapped function's globals. However, the decorator module also uses globals to pass in magical objects such as _decorate_ and _func_. To continue to pass these in, we wrap the generated function in an outer function and use closures.